FDN-3751 Final cut at standalone app
