+++ lib/conf/reference.conf
@@ -1,1 +1,2 @@
 play.modules.enabled += "io.flow.play.CoordinatedShutdownActorReaperModule"
+play.modules.enabled += "io.flow.play.RollbarLifecycleModule"
