--- standalone/src/main/resources/reference.conf
@@ -1,7 +1,0 @@
-// Provides equivalent of io.flow.play.clients.ConfigModule iff lib-play is not on the classpath
-play.modules.enabled += "io.flow.play.standalone.FlowConfigModule"
-
-play.modules.enabled += "io.flow.log.RollbarModule"
-
-// adds shutdown hook to ensure rollbar is closed and flushed
-play.modules.enabled += "io.flow.play.standalone.RollbarLifecycleModule"
